- Updated the firmware to comply with the Pico 2, as opposed to Pico 1. - Modified the SPI requirements to comply with the Pico 2's SPI chip.
