+++ .gitignore
@@ -10,4 +10,4 @@ compile_commands.json
 CTestTestfile.cmake
 _deps
 CMakeUserPresets.json
-Build
+build
