+++ src/spi_comm.c
